refactor: extract beforeunload handler into useBeforeUnload composable
